Problem =========
When user creates a backup configuration with the following playbook, the backup config file additionally
includes string "show running-configuration" in it.This prevents using the config file later to apply in switch.

---
- hosts: leaf
  gather_facts: No
  vars:
    output_dir: "/root/switch_dump/"
  collections:
    -dellemc_networking.os10

  tasks:
  - name: dump switch config
    os10_config:
      backup: yes
      backup_options:
        filename: "{{ inventory_hostname }}.cfg"
        dir_path: "{{ output_dir }}"

Root cause:
In Ansible core module, there is a code to remove "show running-configuration" from the file, if the command
"show running-configuration" was executed.But in out get-config, we have the command to be executed as "show running-config",
which prevents the string from getting removed.

Fix:
Changed the command to be executed from "show running-config" to "show running-configuration"

UT:
Ran the problematic playbook and no issue was observed.